Bug fix on integration functions
There was a major bug that caused Alm to normalise Alm in the Gate case, with the normalisation constant of the Alm(Gauss) case.
This is a critical fix then.

I used the opportunity as well to improve the efficiency of the calculation in the gate case: I now by pass the filter function, which will increase computation speed by several times.