Add CFonly keywords for ajfit models
In order to be able to compute the global likelihood of a model with only the centrigal force, a new switch called do_CFonly must be added into the ajfit .model file. When set to 1 (True), I will use the function decompose_CFonly() to return the expected acoefficients: With current assumption, this means a2_CF only.
When set to 0 (False), the ajfit model will include both the CF term and the Actvity term.