Update of getmodel
Handling a new optional parameter: data_type.
If data_type = file : The behavior is as before (expect a data file in argv[1]). This is the default option
If data_type = range: The expected input in argv[1] is a range + resolution ==> The code will then generate a x-axis according to the user values in order to compute the model.