Partial implementation of the Tabualted_2D priors
A parser of the input / prior table is added in reformat_tabulated_priors(). This allows to identify entries of the form Tabulated_2d(X), with X a name of the input that is correlated to the current entry.
It modifies the table of inputs / priors to be compatible with the way priors_calc::apply_generic_priors() works.