Several bug fixes associated to model_Kallinger2014_Gaussian
a. fully functional set of functions to define initial guess and priors from a power spectrum and for the model using Kallinger+2014 noise model.
b. many minor improvments in the TAMCMC code
c. Preparing the standardisation of the usage of the "model_class" along with the "model_fullname". The standard will be [model_class].[model_fullname]