v1.86.7: Introduction of the getevidence tools  + minor refactoring
This version introduces the getevidence binary tool that allows to recompute the Bayesian evidence from the parallel chains with more controls on the parameters. This version does not include yet the bootstrap method to compute uncertainty of the evidence, but this is planned for rapid implementation.

The version also made some minor changes in the write_evidence() and evidence_calc() functions that allow their call without having to define the global variable of the diagnostic class.